requirements.txt
ccxt>=4.0.0
python-binance==1.0.32
pandas>=2.2.0
numpy>=2.0.0
ta>=0.11.0
aiohttp>=3.9.5
requests>=2.32.0
pyyaml>=6.0.2
websockets>=12.0
cryptography>=42.0.0
python-telegram-bot==13.14
uvicorn>=0.30.0
fastapi>=0.111.0
joblib>=1.3.2
xgboost>=2.1.0
scikit-learn>=1.3.2
tqdm>=4.65.0
jinja2>=3.1.2
aiofiles>=25.1.0
